requirements.txt
Flask==2.2.3
psutil
RPi.GPIO
dht11

